# docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/contracts/upload": {
      "post": {
        "operationId": "ContractController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "contractType": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "summary": "Upload a contract file",
        "tags": [
          "contracts"
        ]
      }
    },
...
    "/api/contracts/{id}": {
      "get": {
        "operationId": "ContractController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the contract"
          }
        },
        "summary": "Get a contract by id",
        "tags": [
          "contracts"
        ]
      },
      "patch": {
        "operationId": "ContractController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContractDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contract updated successfully"
          }
        },
        "summary": "Update a contract",
        "tags": [
          "contracts"
        ]
      },
      "delete": {
        "operationId": "ContractController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract deleted successfully"
          }
        },
        "summary": "Delete a contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/analyze": {
      "post": {
        "operationId": "ContractController_analyzeContract",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Contract analysis completed"
          }
        },
        "summary": "Analyze a contract using AI",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/summary": {
      "get": {
        "operationId": "ContractController_getContractSummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return contract summaries"
          }
        },
        "summary": "Get contract summaries",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/risks": {
      "get": {
        "operationId": "ContractController_getContractRisks",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return contract risks"
          }
        },
        "summary": "Get contract risks",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/analysis": {
      "get": {
        "operationId": "ContractController_getAnalysis",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return analysis data"
          }
        },
        "summary": "Get full contract analysis",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/qna": {
      "get": {
        "operationId": "ContractController_getContractQnA",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return contract Q&A"
          }
        },
        "summary": "Get contract Q&A",
        "tags": [
          "contracts"
        ]
      },
      "post": {
        "operationId": "ContractController_askQuestion",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Question answered successfully"
          }
        },
        "summary": "Ask a question about the contract",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/chat": {
      "post": {
        "operationId": "ContractController_submitChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Chat answered"
          }
        },
        "summary": "Submit a chat question",
        "tags": [
          "contracts"
        ]
      },
      "get": {
        "operationId": "ContractController_getChat",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return chat messages"
          }
        },
        "summary": "Get chat history",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/risk-flags/{riskId}": {
      "patch": {
        "operationId": "ContractController_updateRiskFlag",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "riskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRiskFlagDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Risk flag updated"
          }
        },
        "summary": "Update risk flag status",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/export": {
      "get": {
        "operationId": "ContractController_exportAnalysis",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return analysis export"
          }
        },
        "summary": "Export contract analysis",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/reviews": {
      "get": {
        "operationId": "ContractController_getContractReviews",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return contract reviews"
          }
        },
        "summary": "Get contract reviews",
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/hybrid-ingest": {
      "post": {
        "operationId": "ContractController_hybridIngest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestContractDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
    "/api/contracts/{id}/hybrid-search": {
      "get": {
        "operationId": "ContractController_hybridSearch",
        "parameters": [
          {
            "name": "q",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "contracts"
        ]
      }
    },
...
  },
  "info": {
    "title": "API",
    "description": "API docs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "UpdateContractDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateRiskFlagDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "open",
              "resolved",
              "ignored"
            ]
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "status"
        ]
      },
      "IngestContractDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  }
}